backend/requirements.txt
certifi==2018.4.16
Django==2.0.2
djangorestframework==3.8.2
mkl-fft==1.0.0
mkl-random==1.0.1
numpy==1.14.3
pandas==0.22.0
patsy==0.5.0
python-dateutil==2.7.2
pytz==2018.4
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
statsmodels==0.8.0
wincertstore==0.2
